requirements.txt
fastapi
uvicorn
openai
pyodbc
python-dotenv
urllib3
requests
azure-identity
azure-keyvault-secrets
certifi